+++ torchbearer/torchbearer.py
@@ -13,22 +13,25 @@ from torchbearer.callbacks.printer import Tqdm
 class Model:
     """ Torchbearermodel to wrap base torch model and provide training environment around it
     """
-    def __init__(self, model, optimizer, loss_criterion, metrics=[]):
+    def __init__(self, model, optimizer, criterion=None, metrics=[]):
         """ Create torchbearermodel which wraps a base torchmodel and provides a training environment surrounding it
 
         :param model: The base pytorch model
         :type model: torch.nn.Module
         :param optimizer: The optimizer used for pytorch model weight updates
         :type optimizer: torch.optim.Optimizer
-        :param loss_criterion: The final loss criterion that provides a loss value to the optimizer
-        :type loss_criterion: function
+        :param criterion: The final loss criterion that provides a loss value to the optimizer
+        :type criterion: function or None
         :param metrics: Additional metrics for display and use within callbacks
         :type metrics: list
         """
         super().__init__()
+        if criterion is None:
+            criterion = lambda y_pred, y_true: torch.zeros(y_true.shape[0], device=y_true.device)
+
         self.main_state = {
             torchbearer.MODEL: model,
-            torchbearer.CRITERION: loss_criterion,
+            torchbearer.CRITERION: criterion,
             torchbearer.OPTIMIZER: optimizer,
             torchbearer.DEVICE: 'cpu',
             torchbearer.DATA_TYPE: torch.float32,